Added LBU

+++ Utils.cpp
@@ -1,7 +1,7 @@
 #include "Utils.hpp"
 #include <stdint.h>
 
-std::string Clarkitechture::Utils::toHexStr(uint32_t hex)
+std::string Clarkitechture::Utils::toHexStr(size_t hex)
 {
     std::stringstream hexStream;
     hexStream << std::hex << std::uppercase << hex;
